+++ flake8_trio.py
@@ -65,6 +65,7 @@ Error_codes = {
         "Nurseries should generally be the inner-most context manager."
     ),
     "TRIO112": "Redundant nursery {}, consider replacing with directly awaiting the function call",
+    "TRIO113": "Dangerous `.start_soon()`, process might not run before `__aenter__` exits. Consider replacing with `.start()`.",
 }
 
 
@@ -202,8 +203,9 @@ def has_decorator(decorator_list: List[ast.expr], *names: str):
     return False
 
 
-# matches the full decorator name against fnmatch pattern
-def fnmatch_decorator(decorator_list: List[ast.expr], *patterns: str):
+# matches the fully qualified name against fnmatch pattern
+# used to match decorators and methods to user-supplied patterns
+def fnmatch_qualified_name(name_list: List[ast.expr], *patterns: str):
     def construct_name(expr: ast.expr) -> Optional[str]:
         if isinstance(expr, ast.Call):
             expr = expr.func
@@ -216,12 +218,13 @@ def fnmatch_decorator(decorator_list: List[ast.expr], *patterns: str):
         # See https://peps.python.org/pep-0614/ - we don't handle everything
         return None  # pragma: no cover  # impossible on Python 3.8
 
-    for decorator in decorator_list:
-        qualified_decorator_name = construct_name(decorator)
-        if qualified_decorator_name is None:
+    for name in name_list:
+        qualified_name = construct_name(name)
+        if qualified_name is None:
             continue  # pragma: no cover  # impossible on Python 3.8
         for pattern in patterns:
-            if fnmatch(qualified_decorator_name, pattern.lstrip("@")):
+            # strip leading "@"s for when we're working with decorators
+            if fnmatch(qualified_name, pattern.lstrip("@")):
                 return True
     return False
 
@@ -873,7 +876,7 @@ class Visitor107_108(Flake8TrioVisitor):
         self.set_state(self.default, copy=True)
 
         # disable checks in asynccontextmanagers by saying the function isn't async
-        self.async_function = not fnmatch_decorator(
+        self.async_function = not fnmatch_qualified_name(
             node.decorator_list, *self.options.no_checkpoint_warning_decorators
         )
 
@@ -1149,10 +1152,56 @@ class Visitor107_108(Flake8TrioVisitor):
         self.uncheckpointed_statements = worst_case_shortcut
 
 
+def _get_identifier(node: ast.expr) -> str:
+    if isinstance(node, ast.Name):
+        return node.id
+    if isinstance(node, ast.Attribute):
+        return node.attr
+    return ""  # pragma: no cover
+
+
+class Visitor113(Flake8TrioVisitor):
+    def __init__(self, *args: Any, **kwargs: Any):
+        super().__init__(*args, **kwargs)
+        self.async_function = False
+        self.asynccontextmanager = False
+        self.aenter = False
+
+    def visit_AsyncFunctionDef(self, node: ast.AsyncFunctionDef):
+        outer = self.aenter
+
+        self.aenter = (node.name == "__aenter__" and len(node.args.args) == 1) or any(
+            _get_identifier(d) == "asynccontextmanager" for d in node.decorator_list
+        )
+
+        self.generic_visit(node)
+        self.aenter = outer
+
+    def visit_Yield(self, node: ast.Yield):
+        self.aenter = False
+
+    def visit_Call(self, node: ast.Call) -> None:
+        if (
+            self.aenter
+            and isinstance(node.func, ast.Attribute)
+            and isinstance(node.func.value, ast.Name)
+            # requires that the nursery is named "nursery" for now
+            and node.func.value.id == "nursery"
+            and node.func.attr == "start_soon"
+            and len(node.args) > 0
+            and fnmatch_qualified_name(
+                node.args[:1],
+                *self.options.startable_methods_in_context_manager,
+                *self.options.extend_startable_methods_in_context_manager,
+            )
+        ):
+            self.error("TRIO113", node)
+
+
 class Plugin:
     name = __name__
     version = __version__
-    options: Namespace = Namespace(no_checkpoint_warning_decorators=[])
+    options: Namespace = Namespace()
 
     def __init__(self, tree: ast.AST):
         self._tree = tree
@@ -1181,6 +1230,36 @@ class Plugin:
                 " Decorators can be dotted or not, as well as support * as a wildcard."
                 " For example, ``--no-checkpoint-warning-decorators=app.route,"
                 "mydecorator,mypackage.mydecorators.*``"
+            ),
+        )
+        option_manager.add_option(
+            "--startable-methods-in-context-manager",
+            default="trio.run_process, trio.serve_ssl_over_tcp, trio.serve_tcp, trio.serve_listeners,*.serve",
+            parse_from_config=True,
+            required=False,
+            comma_separated_list=True,
+            help=(
+                "Comma-separated list of method calls to enable TRIO113"
+                " warnings for."
+                " Use if you want to override the default methods the check is enabled for."
+                " Methods can be dotted or not, as well as support * as a wildcard."
+                " For example, ``--startable-methods-in-context-manager=mylib.myfun,"
+                "myfun2,mypackage.myfunlib.*``"
+            ),
+        )
+        option_manager.add_option(
+            "--extend-startable-methods-in-context-manager",
+            default="",
+            parse_from_config=True,
+            required=False,
+            comma_separated_list=True,
+            help=(
+                "Comma-separated list of method calls to enable TRIO113"
+                " warnings for."
+                " Use if you want to extend the list of default methods the check is enabled for."
+                " Methods can be dotted or not, as well as support * as a wildcard."
+                " For example, ``--extend-startable-methods-in-context-manager=mylib.myfun,"
+                "myfun2,mypackage.myfunlib.*``"
             ),
         )